+++ code/searchia-ranking/src/test/java/ir/parsijoo/searchia/RankerTest.java
@@ -100,7 +100,7 @@ class RankerTest {
     void updateRanks() {
         docs.stream().filter(doc -> doc.getId() < 5).forEach(doc -> doc.setRank(1));
         docs.stream().filter(doc -> doc.getId() % 2 == 0).forEach(doc -> doc.setNumberOfMatches(1));
-        List<Long> expectedRanks = List.of(3L, 2L, 3L, 2L, 1L, 0L, 1L, 0L, 1L, 0L, 1L, 0L, 1L, 0L, 1L, 0L, 1L);
+        List<Integer> expectedRanks = List.of(3, 2, 3, 2, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1);
 
         Ranker.updateRanks(docs, Doc::getNumberOfMatches, true);
 
@@ -112,7 +112,7 @@ class RankerTest {
         docs.stream().filter(doc -> doc.getId() < 5).forEach(doc -> doc.setRank(1));
         docs.stream().filter(doc -> doc.getId() % 2 == 0).forEach(doc -> doc.setNumberOfMatches(1));
         docs.stream().filter(doc -> doc.getId() > 10).forEach(doc -> doc.setNumberOfMatches(Integer.MAX_VALUE));
-        List<Long> expectedRanks = List.of(4L, 3L, 4L, 3L, 2L, 1L, 2L, 1L, 2L, 1L, 0L, 0L, 0L, 0L, 0L, 0L, 0L);
+        List<Integer> expectedRanks = List.of(4, 3, 4, 3, 2, 1, 2, 1, 2, 1, 0, 0, 0, 0, 0, 0, 0);
 
         Ranker.updateRanks(docs, Doc::getNumberOfMatches, true);
